% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_frag.R
\name{add_child}
\alias{add_child}
\title{Add a child fragment}
\usage{
add_child(x, tag, contents, ..., .attr = list(...))
}
\arguments{
\item{x}{an xml_fragment}

\item{tag}{character, the name of the tag}

\item{contents}{character, xml_fragment or list of xml_fragment or character strings}

\item{...}{additional attributes to add to the tag}

\item{.attr}{a list of additional attributes to add to the tag, overrides the \code{...} argument}
}
\value{
an xml_fragment with the new child added
}
\description{
Adds a child fragment to an existing xml_fragment
}
\seealso{
Other xml_fragment: 
\code{\link{append_frag}()},
\code{\link{frag}()}
}
\concept{xml_fragment}
