% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_frag.R
\name{append_frag}
\alias{append_frag}
\title{Add a fragment with a tag to an existing xml_fragment}
\usage{
append_frag(x, tag, contents, ..., .attr = list(...))
}
\arguments{
\item{x}{an \code{xml_fragment}}

\item{tag}{character, the name of the tag (single value)}

\item{contents}{character, xml_fragment or list of xml_fragment or character strings}

\item{...}{additional attributes to add to the tag}

\item{.attr}{a list of additional attributes to add to the tag, override the \code{...} argument}
}
\value{
an \code{xml_fragment} with the new tag added
}
\description{
\code{append_frag} add a tag with supplied \code{contents} to an \code{xml_fragment}. The \code{tag} argument
is the name of the tag, and \code{contents} is the contents of the tag, which can be
an xml_fragment, a character string, or a list of xml_fragment or character strings.
}
\seealso{
Other xml_fragment: 
\code{\link{add_child}()},
\code{\link{frag}()}
}
\concept{xml_fragment}
