% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_fragment.R
\name{as.character.xml_fragment}
\alias{as.character.xml_fragment}
\alias{as.character.xml_doc}
\title{Turn an xml_fragment into a character}
\usage{
\method{as.character}{xml_fragment}(x, use_prolog = FALSE, ...)

\method{as.character}{xml_doc}(x, use_prolog = TRUE, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{use_prolog}{if \code{TRUE} the xml prolog with be included.
To suppress the prolog string either remove set \code{use_prolog = FALSE}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{character} with the xml representation of the fragment.
}
\description{
This function turns an \code{xml_fragment} into a character string, using
a performant c++ implementation.
}
\seealso{
Other xml_fragment: 
\code{\link{add_child_fragment}()},
\code{\link{as_frag}()},
\code{\link{as_xml_nodeset}()},
\code{\link{data_frag}()},
\code{\link{frag}()},
\code{\link{xml_fragment}()}
}
\concept{xml_fragment}
