% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_frag.R
\name{as_frag}
\alias{as_frag}
\title{Convert a list to an xml fragment}
\usage{
as_frag(x, ..., .attr = list(...))
}
\arguments{
\item{x}{named list that will be transformed into a fragment}

\item{...}{optional attributes to be set on the parent of the fragment}
}
\value{
\code{\link[=xml_fragment]{xml_fragment()}} object as if specified with \code{\link[=frag]{frag()}}.
}
\description{
As frag is a helper function to convert a named list to an xml fragment, it
transforms all values to character, and recursively transforms nested lists.
\code{as_frag} can be used for flexible list created xml fragments, names of a
list turn into tags, and values into text nodes.
}
\seealso{
Other xml_fragment: 
\code{\link{add_child_fragment}()},
\code{\link{as.character.xml_fragment}()},
\code{\link{as_xml_nodeset}()},
\code{\link{data_frag}()},
\code{\link{frag}()},
\code{\link{xml_fragment}()}
}
\concept{xml_fragment}
