% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_fragment.R
\name{as_xml_nodeset}
\alias{as_xml_nodeset}
\title{Transforms an xml_fragment into an xml_nodeset}
\usage{
as_xml_nodeset(x, ...)
}
\arguments{
\item{x}{an object created with \code{\link[=xml_fragment]{xml_fragment()}}}

\item{...}{reserved for future use}
}
\description{
Using the \code{xml2} package, this function transforms an \code{xml_fragment} into an \code{xml_nodeset}
}
\seealso{
Other xml_fragment: 
\code{\link{add_child_fragment}()},
\code{\link{as.character.xml_fragment}()},
\code{\link{as_frag}()},
\code{\link{data_frag}()},
\code{\link{frag}()},
\code{\link{xml_fragment}()}

Other xml2: 
\code{\link{list_as_xml_document}()},
\code{\link{list_as_xml_string}()}
}
\concept{xml2}
\concept{xml_fragment}
