% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_xml_node.R
\name{list_to_xml_node}
\alias{list_to_xml_node}
\title{Convert a list to an xml node}
\usage{
list_to_xml_node(x, ...)
}
\arguments{
\item{x}{a list as returned by \code{\link[xml2:as_list]{xml2::as_list()}}}

\item{...}{reserved for future use}
}
\value{
an \code{xml2::xml_document}
}
\description{
\code{list_to_xml_node} is fast and efficient way to convert a list to an \code{xml2::xml_node}.
}
\details{
\code{list_to_xml_node} is a much faster implementation of \code{xml2::as_xml_document.list()} method.
It writes the xml directly to a string and then reads it back into an \code{xml2::xml_node}.
Note that it returns the root node of the document, not the document itself.
It therefore can be used to create fragments or single nodes.

The function can be used in tandem with \code{\link[xml2:as_list]{xml2::as_list()}} to convert R data structures.
}
\seealso{
Other xml2: 
\code{\link{list_to_xml_string}()}
}
\concept{xml2}
