% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_xml_string.R
\name{list_to_xml_string}
\alias{list_to_xml_string}
\title{Convert a list to an xml string}
\usage{
list_to_xml_string(x, ...)
}
\arguments{
\item{x}{a list as returned by \code{\link[xml2:as_list]{xml2::as_list()}}}

\item{...}{reserved for future use}
}
\description{
\code{list_to_xml_string} is fast and efficient way to convert a list to an xml string.
}
\details{
The list format is identical to the format returned by \code{\link[xml2:as_list]{xml2::as_list()}} function,
but much, much faster. This function allows for easy conversion of R data structures
into xml format by creating the list structures in R and then converting them to xml.

The function can be used in tandem with \code{\link[xml2:as_list]{xml2::as_list()}} to convert R data structures
}
\seealso{
Other xml2: 
\code{\link{list_to_xml_node}()}
}
\concept{xml2}
