% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlbuilder.R
\name{xmlbuilder}
\alias{xmlbuilder}
\title{Create an fast feed-forward XML builder}
\usage{
xmlbuilder(use_prolog = TRUE, single_root_warn = TRUE)
}
\arguments{
\item{use_prolog}{logical. Should the XML prolog be included in the output?
Default is \code{TRUE}, which generate an UTF-8 xml prolog.
Set to \code{FALSE} if you want to generate an xml fragment or manually prepend the prolog.}

\item{single_root_warn}{logical. Should a warning be issued if the XML document has multiple root elements?
Set to \code{FALSE} to suppress when creating multiple xml fragments.}
}
\value{
An object of class `xmlbuilder
}
\description{
This function creates an XML builder that allows you to create XML documents in a feed-forward manner.
\code{xmlbuilder} returns an object that has methods to create XML elements, text nodes, comments, and more.
}
\details{
\itemize{
\item \verb{$start(tag, ...)} (or \verb{$start_element}) starts an element with a given tag and attributes.
\item \verb{$end()} (or \verb{$end_element}) ends the current element.
\item \verb{$element(tag, text, ...)} creates an element with a given tag, text, and attributes.
\item \verb{$text(text)} creates a text node.
\item \verb{$comment(comment)} creates a comment node.
\item \verb{$xml_string()} returns the XML document or fragments(s) as a character vector.
}
}
\examples{
example/xmlbuilder.R
}
