% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frag.R
\name{data_frag}
\alias{data_frag}
\title{Create an xml_fragment from a data.frame}
\usage{
data_frag(df, row_tags = "row")
}
\arguments{
\item{df}{data frame that will be stored as set of xml elements}

\item{row_tags}{\code{character} the tag name that is used for each row. Note that
this can be a single value or a vector of length equal to the number of rows in the data.frame.}
}
\value{
\code{\link[=xml_fragment]{xml_fragment()}} object
}
\description{
Create a \code{\link[=xml_fragment]{xml_fragment()}} from a data.frame, in which each row is a set of xml elements (columns).
}
\examples{
persons <- data.frame(
  name = c("John Doe", "Jane Doe"),
  age = c(30, 25),
  stringsAsFactors = FALSE
)

df <- data_frag(persons, row_tag = "person")
print(df)

tag("study", date = "today") / df
}
\seealso{
Other xml_fragment: 
\code{\link{add_child_fragment}()},
\code{\link{as_xml_nodeset}()},
\code{\link{frag}()}
}
\concept{xml_fragment}
